requirements.txt
confluent-kafka
requests
six
pandas
numpy
sklearn
imblearn
matplotlib
seaborn 
xgboost 
statistics
jsonschema==3.2.0
pymongo[srv]==4.1.1
dnspython
pyyaml
from_root
dill
fastapi
uvicorn
-e .
